Added Automatic Token Retrieval option.
This option creates a new despondency that is NOT included with the source code and will NOT be publicly available.
This will prevent users from being able to compile the project themselves, they will have to figure out how to recreate the required dependency.

Reason:
The method used to retrieve and decrypt the token is the same as what malware can use, in order to prevent potential abuse it will not be included.